+++ CHANGES.md
@@ -5,3 +5,5 @@ Next release
 - Use `*_opt` naming convention for the functions that return an option
   instead of an exception: `Py.import_opt`, `Py.Object.find_opt`,...
 - of_seq/to_seq converters
+- [*] get_attr/get_attr_string now returns option type
+- Indexing operators (for OCaml 4.06.0 and above) defined in Pyops